Split directions into separate module
