+++ js/src/index.js
@@ -16,5 +16,6 @@ export * from './Figure';
 export * from './Heatmap';
 export * from './Marker';
 export * from './GeoJson';
+export * from './Directions';
 
 export { version } from '../package.json';
