+++ js/src/jupyter-gmaps.js
@@ -30,61 +30,6 @@ const ConfigurationMixin = {
 
 // Views
 
-
-export const DirectionsLayerView = GMapsLayerView.extend({
-    render() {
-        const rendererOptions = { map: this.mapView.map }
-
-        this.directionsDisplay = new google.maps.DirectionsRenderer(rendererOptions)
-
-        const modelData = this.model.get("data");
-
-        const request = {
-            origin: this.getOrigin(modelData),
-            destination: this.getDestination(modelData),
-            waypoints: this.getWaypoints(modelData),
-            travelMode: google.maps.TravelMode.DRIVING
-        };
-
-        const directionsService = new google.maps.DirectionsService();
-
-        directionsService.route(request, (response, status) => {
-            // print to the browser console (mostly for debugging)
-            console.log(`Direction service returned: ${status}`) ;
-            // set a flag in the model
-            this.model.set("layer_status", status) ;
-            this.touch() ; // push `layer_status` changes to the model
-            if (status == google.maps.DirectionsStatus.OK) {
-                this.response = this.directionsDisplay ;
-                this.directionsDisplay.setDirections(response);
-            }
-        });
-    },
-
-
-    addToMapView(mapView) { },
-
-    getOrigin(modelData) {
-        const [lat, lng] = _.first(modelData)
-        return new google.maps.LatLng(lat, lng)
-    },
-
-    getDestination(modelData) {
-        const [lat, lng] = _.last(modelData)
-        return new google.maps.LatLng(lat, lng)
-    },
-
-    getWaypoints(modelData) {
-        const withoutFirst = _.tail(modelData)
-        const withoutLast = _.initial(withoutFirst)
-        const dataAsGoogle = withoutLast.map(([lat, lng]) => {
-            return {location: new google.maps.LatLng(lat, lng)}
-        })
-        return dataAsGoogle
-    }
-})
-
-
 export const PlainmapView = widgets.DOMWidgetView.extend({
     render() {
         this.loadConfiguration();
@@ -155,13 +100,6 @@ _.extend(PlainmapView.prototype, ConfigurationMixin);
 
 
 // Models
-
-export const DirectionsLayerModel = GMapsLayerModel.extend({
-    defaults: _.extend({}, GMapsLayerModel.prototype.defaults, {
-        _view_name: "DirectionsLayerView",
-        _model_name: "DirectionsLayerModel"
-    })
-});
 
 export const PlainmapModel = widgets.DOMWidgetModel.extend({
     defaults: _.extend({}, widgets.DOMWidgetModel.prototype.defaults, {
